requirements.txt
aiofiles==23.2.1
aiogram==3.17.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiohttp-sse-client==0.2.1
aiosignal==1.3.2
alembic==1.13.3
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.10.4
attrs==23.2.0
bitarray==2.9.3
certifi==2023.11.17
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
et_xmlfile==2.0.0
frozenlist==1.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytonconnect==0.3.2
pytoniq-core==0.1.44
pytz==2024.2
requests==2.31.0
setuptools==70.3.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.43
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
urllib3==2.2.3
x25519==0.0.2
yarl==1.9.11
